decouple `ApplicationSupervisor` from `UIApplicationDelegate` This enables plugins to be registered and referenced independently from the UIApplicationDelegate lifecycle.
Previously, referencing a plugin API was dependent on calling
`ApplicationSupervisor.sharedInstance` which was computed by calling
`UIApplication.shared.delegate as! ApplicationSupervisor`. This is
problematic since we do not control the instantiation of the `delegate`
property and it makes it unsafe to reference feature APIs from a
background queue.
